Implement DetailPage and Base Detail Pages for Modules, Observations, Visits and Sites:
  1. Cohérence architecturale : Nous avons appliqué un pattern uniforme à toutes les classes de
  détail, où chaque composant suit la même structure avec une séparation claire entre :
    - L'adaptateur (ConsumerStatefulWidget) qui accède à Riverpod
    - La classe base (DetailPage) qui implémente la logique métier
  2. Meilleure gestion des dépendances : Le paramètre ref est maintenant explicitement passé aux
  classes de base, ce qui :
    - Rend la dépendance à Riverpod visible et traçable
    - Élimine les "aspects magiques" où le code fonctionnait sans qu'on comprenne pourquoi
    - Facilite les tests unitaires en permettant d'injecter des mocks
  3. Réduction du couplage : Les classes de base ne sont plus dépendantes de l'implémentation
  spécifique de Riverpod, ce qui permet :
    - De changer plus facilement le système de gestion d'état si nécessaire
    - D'avoir un code plus modulaire et réutilisable
  4. Gestion des cycles de vie améliorée : Avec ConsumerStatefulWidget au lieu de ConsumerWidget :
    - Nous pouvons contrôler précisément quand l'injection des dépendances se produit
    - L'initialisation et le nettoyage des ressources sont gérés de manière appropriée
  5. Retours de fonctions explicites : La correction des avertissements "unused result" assure que :
    - Les fonctions retournent explicitement des valeurs quand nécessaire
    - Le code est plus prévisible et moins sujet aux bugs subtils
    - Les futures modifications seront plus faciles car le comportement attendu est clair
  6. Imports optimisés : La suppression des imports inutilisés :
    - Réduit la confusion sur les dépendances réelles du code
    - Améliore légèrement les performances de compilation
    - Rend le code plus propre et plus maintenable
  7. Pattern d'injection de dépendances standardisé : L'utilisation de
  WidgetsBinding.instance.addPostFrameCallback pour l'injection assure que :
    - Les dépendances sont injectées une fois que le widget est complètement monté
    - Les erreurs d'initialisation sont capturées et gérées de manière élégante
    - Le pattern est cohérent dans toute l'application
